Rename boomark to bookmark, move ratinghistory
