+++ react-frontend/src/components/layout/Navigation.jsx
@@ -19,7 +19,7 @@ const Navigation = () => {
           <Nav.Link href="#home">Home</Nav.Link>
           <Nav.Link href="#features">Features</Nav.Link>
           <Nav.Link href="#pricing">Pricing</Nav.Link>
-          <Nav.Link href="#bookmark">Bookmark</Nav.Link>
+          <Nav.Link href="bookmark">Bookmark</Nav.Link>
           <Nav.Link href="ratinghistory">RatingHistory</Nav.Link>
         </Nav>
       </Container>
