+++ react-frontend/src/components/privateViews/Bookmark.jsx
@@ -1,0 +1,5 @@
+const Bookmark = () => {
+  return <div></div>;
+};
+
+export default Bookmark;
