--- react-frontend/src/components/privateViews/Boomark.jsx
@@ -1,5 +1,0 @@
-const Boomark = () => {
-  return <div></div>;
-};
-
-export default Boomark;
