+++ react-frontend/src/main.jsx
@@ -3,7 +3,8 @@ import ReactDOM from "react-dom/client";
 
 import App from "./App.jsx";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
-import RatingHistory from "./components/views/RatingHistory.jsx";
+import RatingHistory from "./components/privateViews/RatingHistory.jsx";
+import { Bookmark } from "react-bootstrap-icons";
 
 const router = createBrowserRouter([
   {
@@ -13,6 +14,10 @@ const router = createBrowserRouter([
   {
     path: "/ratinghistory",
     element: <RatingHistory />,
+  },
+  {
+    path: "/bookmark",
+    element: <Bookmark />,
   },
 ]);
